+++ packages/create-hydrogen-app/CHANGELOG.md
@@ -8,6 +8,7 @@ and adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 ## Unreleased
 
 - Wrap instances of `<Money>` in client components due to render prop usage
+- Eliminate use of `Link` client re-exports
 
 ## 0.6.1 - 2021-11-08