+++ packages/create-hydrogen-app/CHANGELOG.md
@@ -5,7 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-<!-- ## Unreleased -->
+## Unreleased
+
+- Wrap instances of `<Money>` in client components due to render prop usage
+- Eliminate use of `Link` client re-exports
 
 ## 0.6.1 - 2021-11-08